generate reborrow constraints at type check time
