+++ src/librustc_mir/borrow_check/nll/region_infer/mod.rs
@@ -333,17 +333,6 @@ impl<'tcx> RegionInferenceContext<'tcx> {
         self.liveness_constraints.add_element(v, element)
     }
 
-    /// Indicates that the region variable `sup` must outlive `sub` is live at the point `point`.
-    pub(super) fn add_outlives(&mut self, locations: Locations, sup: RegionVid, sub: RegionVid) {
-        assert!(self.inferred_values.is_none(), "values already inferred");
-        self.constraints.push(OutlivesConstraint {
-            locations,
-            sup,
-            sub,
-            next: None,
-        })
-    }
-
     /// Perform region inference and report errors if we see any
     /// unsatisfiable constraints. If this is a closure, returns the
     /// region requirements to propagate to our creator, if any.
